Share code between FormInput and MasedFormInput
No more duplication!
